+++ babble/src/main/java/io/mosaicnetworks/babble/configure/JoinGroupFragment.java
@@ -222,7 +222,7 @@ public class JoinGroupFragment extends Fragment implements ResponseListener {
         }
 
         mLoadingDialog.dismiss();
-        mListener.onJoined(mMoniker);
+        mListener.baseOnJoined(mMoniker);
     }
 
     @Override